+++ src/Ext/House/Body.h
@@ -66,6 +66,8 @@ public:
 
 		int ForceEnemyIndex;
 
+		int BattlePoints;
+
 		ExtData(HouseClass* OwnerObject) : Extension<HouseClass>(OwnerObject)
 			, PowerPlantEnhancers {}
 			, OwnedLimboDeliveredBuildings {}
@@ -94,6 +96,7 @@ public:
 			, SuspendedEMPulseSWs {}
 			, SuperExts(SuperWeaponTypeClass::Array.Count)
 			, ForceEnemyIndex(-1)
+			, BattlePoints(0)
 		{ }
 
 		bool OwnsLimboDeliveredBuilding(BuildingClass* pBuilding);
@@ -114,6 +117,10 @@ public:
 		virtual void InvalidatePointer(void* ptr, bool bRemoved) override;
 
 		void UpdateVehicleProduction();
+
+		void UpdateBattlePoints(int modifier);
+		bool AreBattlePointsEnabled();
+		int CalculateBattlePoints(TechnoClass* pTechno);
 
 		virtual void LoadFromStream(PhobosStreamReader& Stm) override;
 		virtual void SaveToStream(PhobosStreamWriter& Stm) override;