+++ src/Ext/Rules/Body.h
@@ -238,6 +238,10 @@ public:
 		int TintColorForceShield;
 		int TintColorBerserk;
 
+		Nullable<bool> BattlePoints;
+		Nullable<int> BattlePoints_DefaultValue;
+		Nullable<int> BattlePoints_DefaultFriendlyValue;
+
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
 			, HarvesterDumpAmount { 0.0f }
@@ -420,6 +424,10 @@ public:
 
 			, AttackMove_IgnoreWeaponCheck { false }
 			, AttackMove_StopWhenTargetAcquired { }
+
+			, BattlePoints {}
+			, BattlePoints_DefaultValue {}
+			, BattlePoints_DefaultFriendlyValue {}
 		{ }
 
 		virtual ~ExtData() = default;